dom/ipc/tests/test_force_oop_iframe.html

<!DOCTYPE HTML>
<html>
<head>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/ChromeTask.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>

<script type="application/javascript">
"use strict";
/* eslint-env mozilla/frame-script */

add_task(async function() {
  await SpecialPowers.pushPrefEnv({"set": [["fission.oopif.attribute", true],
                                           ["dom.ipc.processCount", 10000],
                                           // Temporarily allow use of eval() in ChromeTask.js.
                                           ["security.allow_eval_with_system_principal", true]]});

  // This iframe should be loaded out of process. Unfortunately as of the time
  // of this test's creation, many different events which we could use to detect
  // this load have not been implemented yet.
  let contentCreated = ChromeTask.spawn(null, async function() {
    let wgp = await new Promise(resolve => {
      function observer(parent) {
        info("WGP with origin: " + parent.documentPrincipal.origin);
        if (parent.documentPrincipal.origin !== "http://mochi.test:8888") {
          return;
        }

        Services.obs.removeObserver(observer, "window-global-created");
        resolve(parent);
      }
      Services.obs.addObserver(observer, "window-global-created");
    });

    is(wgp.isInProcess, false, "not in-process");
    ok(wgp.rootFrameLoader, "Has frameloader");
    ok(wgp.documentPrincipal, "Has document principal");
  });

  var iframe = document.createElement("iframe");
  iframe.setAttribute("fission", "true");
  iframe.setAttribute("src", "file_dummy.html");
  document.body.appendChild(iframe);

  // Check that this isn't loaded in-process, or using a nested remoteTab object.
  let frameLoader = SpecialPowers.wrap(iframe).frameLoader;
  is(frameLoader.docShell, null);
  is(frameLoader.remoteTab, null);
  let browsingContext = frameLoader.browsingContext;
  isnot(browsingContext, null);
  is(browsingContext.docShell, null);

  await contentCreated;

  ok(frameLoader.browsingContext, "Has BrowsingContext");
  ok(frameLoader.browsingContext.parent, "BrowsingContext has parent");

  let wgc = SpecialPowers.wrap(window).getWindowGlobalChild();
  ok(wgc, "we have a WindowGlobalChild");
  is(SpecialPowers.unwrap(wgc.browsingContext),
     SpecialPowers.unwrap(frameLoader.browsingContext.parent),
     "BrowsingContext matches");
});

</script>
</body>
</html>
